feat: initialize project
Add a basic CLI built with Deno that detects
GitHub repositories and run plugins that
build pipeline pieces from detected files
in the project.

A first builtin plugin to validate the tool
checks if there is any Python code and
suggests to run the flake8 linter against
the code base.