feat: rewrite js and python plugins
Introduce technology plugins, they're
focused on learning about what languages,
frameworks and types of files are used
in the project. They don't care about
the CI platform or what the CI output
should be.

The CI files will be generated through
templates that received data introspected
with the technology plugins.