feat: add .gitignore validation for glob
Today the glob pattern is not using the "exclude" parameter,
which results in the analysis of all files that match the glob,
even inside unwanted directories.
This commit adds a function that checks if the .gitignore file exists
and uses its contents to pass it as a list in the glob's exclude parameter.

ref:4