feat: remove prompt from python stack
This commit removes the prompt to the user
asking what Python version the project uses
if it can't be detected.

To achieve this, this commit also adds a
function in the context to emit errors from
stack plugins and allow an introspection to
return undefined. The undefined return should
be interpreted as an failure to introspect
the stack.

ref #11