refactor: separate project in core and cli
Split the project in two main parts: the
core and the cli.

The core module contains plugin implementations
and the builtin templates, as well as types
and interfaces definitions required to use
the core code.

The cli module contains the code to build the
CLI using the code and builtin templates from
the core.