feat: introduce strict mode (#82)
Add a flag to the context object called strict,
that toggles the strict mode. When the strict mode
is on, pipelinit is more demanding and may refuse
to generate a pipeline before something is changed
in the project. When the strict mode is off,
pipelinit won't err, but may generate less accurate
or less complete pipelines.

This is useful to allow distinct execution modes,
based in the user environment: can they easily
change the project and try again? If the aswner
is yes, then the strict mode is better.