fix(core): quote version used in YAML (#84)
Using a semantic version without quotes in the CI YAML files isn't correct. Since YAML supports floating point numbers, a version like 3.10 could be interpreted as 3.1; 3.10 and any other version with a minor or patch with double digits are ok.
Semantic version values should be handled as strings, not as numbers.