fix: adjust javascript lint template and python dependencies search
The javascript lint template in some conditional added some uncessary spaces
that ended up broken the pipeline on Github.
Adjusted the python dependecies search to only see a package manager file
in the project root path with that, we can avoid a situation where Pipelinit
detects a dependency, but when running the pipeline  it can't install it
leaving the pipeline non-functional.

Resolves: https://github.com/pipelinit/pipelinit-cli/issues/88