refactor:  remove hard-coded stack versions
As said on the issue https://github.com/pipelinit/pipelinit-cli/issues/152

Today on Pipelinit the definition of a determined stack the latest release is found inside the code
this doesn't make it easy to update some new minor version with bug fixes.

This commit proposes to remove from Pipelinit the responsibility of knowing the "minor" version
leaving this task to the CI Platform.

Resolves: #152