fix: adjust node_modules exclude
By default when Pipelinit introspects a repository
It uses the repo `.gitignore` to exclude unecessary/static files
which slowdown the cli run, this works for most of the scenarios
however in some cases when the introspected repo has multiples
`node_modules` this may not happen.

This commit adds a manual rule in the globExclude option
to filter ANY `node_modules` directory and any `hidden directory`.

NOTE: Only hidden directories are affected. Pipelinit MUST
check some hidden files in is introspection like `.nvmrc` and
`.eslintrc`.